requirements.txt
opencv-python
numpy
ultralytics
easyocr
flask
scipy
pandas
deep-sort-realtime
python-dotenv
